requirements.txt
streamlit
requests
packaging
python-dotenv
